refactor: fix corrupted import paths and clean up remaining scratchpad references
