+++ l10n/bundle.l10n.json
@@ -239,7 +239,7 @@
   "Click here to retry": "Click here to retry",
   "Click here to update credentials": "Click here to update credentials",
   "Click to learn how to connect": "Click to learn how to connect",
-  "Click to learn how to connect a database for the query playground": "Click to learn how to connect a database for the query playground",
+  "Click to learn how to connect a database for the Query Playground": "Click to learn how to connect a database for the Query Playground",
   "Click to view resource": "Click to view resource",
   "Close tips": "Close tips",
   "Cluster metadata not initialized. Client may not be properly connected.": "Cluster metadata not initialized. Client may not be properly connected.",
@@ -367,9 +367,6 @@
   "DocumentDB for VS Code is not signed in to Azure": "DocumentDB for VS Code is not signed in to Azure",
   "DocumentDB Local": "DocumentDB Local",
   "DocumentDB Performance Tips": "DocumentDB Performance Tips",
-  "Query Playground": "Query Playground",
-  "Query Playground connected to {0}": "Query Playground connected to {0}",
-  "Query Playground connected to {0}/{1}": "Query Playground connected to {0}/{1}",
   "Documents": "Documents",
   "Documents Examined": "Documents Examined",
   "Documents Returned": "Documents Returned",
@@ -801,6 +798,10 @@
   "Query Performance Analysis": "Query Performance Analysis",
   "Query Performance Insight": "Query Performance Insight",
   "Query Plan Summary": "Query Plan Summary",
+  "Query Playground": "Query Playground",
+  "Query Playground connected to {0}": "Query Playground connected to {0}",
+  "Query Playground connected to {0}/{1}": "Query Playground connected to {0}/{1}",
+  "Query playground execution failed: {0}": "Query playground execution failed: {0}",
   "Quick Actions": "Quick Actions",
   "Recommendation: Create Index": "Recommendation: Create Index",
   "Recommendation: Drop Index": "Recommendation: Drop Index",
@@ -853,7 +854,6 @@
   "Saving \"{path}\" will update the entity \"{name}\" to the cloud.": "Saving \"{path}\" will update the entity \"{name}\" to the cloud.",
   "Saving credentials for \"{clusterName}\"…": "Saving credentials for \"{clusterName}\"…",
   "Schema scan complete: {0} fields discovered in \"{1}\".": "Schema scan complete: {0} fields discovered in \"{1}\".",
-  "Query playground execution failed: {0}": "Query playground execution failed: {0}",
   "See output for more details.": "See output for more details.",
   "Select {0}": "Select {0}",
   "Select a location for new resources.": "Select a location for new resources.",
