+++ src/commands/playground/runAll.ts
@@ -6,8 +6,8 @@
 import { type IActionContext } from '@microsoft/vscode-azext-utils';
 import * as l10n from '@vscode/l10n';
 import * as vscode from 'vscode';
-import { PlaygroundService } from '../../documentd./playground/PlaygroundService';
-import { PLAYGROUND_LANGUAGE_ID } from '../../documentd./playground/constants';
+import { PlaygroundService } from '../../documentdb/playground/PlaygroundService';
+import { PLAYGROUND_LANGUAGE_ID } from '../../documentdb/playground/constants';
 import { executePlaygroundCode } from './executePlaygroundCode';
 
 /**
